requirements.txt
# Core Libraries
pandas>=1.5.0
numpy>=1.24.0
ib_insync>=0.9.70

# Technical Analysis
TA-Lib>=0.4.0
pandas_ta>=0.3.14

# Backtesting
backtrader>=1.9.76

# Data Storage
h5py>=3.7.0
tables>=3.8.0

# Networking & APIs
requests>=2.28.0
websocket-client>=1.4.0

# Scientific Computing
scipy>=1.10.0
scikit-learn>=1.3.0

# Monitoring & Logging
flask>=2.2.0
plotly>=5.11.0
dash>=2.10.0

# Alerts & Notifications
python-telegram-bot>=20.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0

# Database
sqlalchemy>=2.0.0
